# backend/api/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/auth/login": {
            "post": {
                "operationId": "AuthController_credentailLogin",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CredentialLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/signup": {
            "post": {
                "operationId": "AuthController_credentailSignUp",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CredentailRegister"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Unable to create user"
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "operationId": "AuthController_logout",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
        "/auth/refresh": {
            "get": {
                "operationId": "AuthController_refresh",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/auth/personal-details": {
            "get": {
                "operationId": "AuthController_me",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth"
                ]
            }
        },
...
        "/auth/google/login": {
            "get": {
                "operationId": "GoogleController_googleAuthFrontend",
                "parameters": [
                    {
                        "name": "idToken",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "google"
                ]
            }
        },
        "/auth/otp/generate-otp": {
            "post": {
                "operationId": "OtpController_generateOtp",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "otp"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/auth/otp/verify-otp": {
            "post": {
                "operationId": "OtpController_verifyOtp",
                "parameters": [
                    {
                        "name": "otp",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "otp"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/exams": {
            "post": {
                "operationId": "ExamsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateExamDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "exams"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "get": {
                "operationId": "ExamsController_findAll",
                "parameters": [
                    {
                        "name": "from",
                        "required": false,
                        "in": "query",
                        "description": "Filter by this date",
                        "schema": {
                            "format": "2021-01-01T11:59:59.999Z",
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "required": false,
                        "in": "query",
                        "description": "Filter by this date",
                        "schema": {
                            "format": "2021-01-01T11:59:59.999Z",
                            "type": "string"
                        }
                    },
                    {
                        "name": "isDeleted",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "contains",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamEntity"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "exams"
                ]
            }
        },
        "/exams/{id}": {
            "get": {
                "operationId": "ExamsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityWithQuestions"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "exams"
                ]
            },
            "patch": {
                "operationId": "ExamsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateExamDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "exams"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "delete": {
                "operationId": "ExamsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "minimum": 1,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "exams"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "NestJS API 0.0.1}",
        "description": "The NestJS API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "http",
                "in": "header",
                "name": "auth-cookie",
                "description": "Cookies are set automatically. No need to add.",
                "scheme": "bearer"
            },
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://localhost:3000/auth/google/backend-login",
                        "scopes": {
                            "read:users": "read users",
                            "write:users": "modify users"
                        }
                    }
                }
            }
        },
        "schemas": {
            "UserEntity": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER"
                        ],
                        "default": "USER"
                    },
                    "id": {
                        "type": "number"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "emailVerified": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "role",
                    "id",
                    "email",
                    "firstName",
                    "lastName",
                    "image",
                    "emailVerified"
                ]
            },
            "ValidationErrorDto": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message",
                    "error"
                ]
            },
            "CredentialLogin": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "pattern": "/[@$!%*#?&]/",
                        "example": "Password@123",
                        "description": "password must contain atleast 1 special character, 1 upper case and 1 number"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "password",
                    "email"
                ]
            },
            "CredentailRegister": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "pattern": "/[@$!%*#?&]/",
                        "example": "Password@123",
                        "description": "password must contain atleast 1 special character, 1 upper case and 1 number"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "object",
                        "default": "USER"
                    }
                },
                "required": [
                    "password",
                    "email",
                    "role"
                ]
            },
            "CreateExamDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ]
            },
            "ExamEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "TagEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "TrueFalseEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "type": {
                        "type": "object",
                        "default": "TRUE_FALSE"
                    },
                    "question": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "answer": {
                        "type": "boolean"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagEntity"
                        }
                    }
                },
                "required": [
                    "id",
                    "type",
                    "question",
                    "description",
                    "answer",
                    "tags"
                ]
            },
            "MultipleChoiceEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "question": {
                        "type": "string"
                    },
                    "type": {
                        "type": "object",
                        "default": "MULTIPLE_CHOICE"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "answer": {
                        "type": "number"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagEntity"
                        }
                    }
                },
                "required": [
                    "id",
                    "question",
                    "type",
                    "description",
                    "options",
                    "answer",
                    "tags"
                ]
            },
            "ShortAnswerEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "question": {
                        "type": "string"
                    },
                    "type": {
                        "type": "object",
                        "default": "SHORT_ANSWER"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "answer": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagEntity"
                        }
                    }
                },
                "required": [
                    "id",
                    "question",
                    "type",
                    "description",
                    "answer",
                    "tags"
                ]
            },
            "EntityWithQuestions": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "trueFalse": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TrueFalseEntity"
                        }
                    },
                    "multipleChoice": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MultipleChoiceEntity"
                        }
                    },
                    "shortAnswer": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShortAnswerEntity"
                        }
                    },
                    "multipleSelect": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MultipleChoiceEntity"
                        }
                    }
                },
                "required": [
                    "id",
                    "name",
                    "createdAt",
                    "updatedAt",
                    "trueFalse",
                    "multipleChoice",
                    "shortAnswer",
                    "multipleSelect"
                ]
            },
            "UpdateExamDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}